Error Handling Post Put Delete

+++ spring-boot-backend-apirest/src/main/java/com/remotejob/springboot/backend/apirets/controllers/ClienteRestController.java
@@ -1,9 +1,13 @@
 package com.remotejob.springboot.backend.apirets.controllers;
 
 import java.util.List;
+import java.util.Map;
+import java.util.HashMap;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -32,31 +36,90 @@ public class ClienteRestController {
 	}
 	
 	@GetMapping("/clientes/{id}")
-	public Cliente show(@PathVariable Long id) {
-		return clienteService.findById(id);
+	public ResponseEntity<?> show(@PathVariable Long id) {
+		Cliente cliente = null;
+		Map<String, Object> response = new HashMap<>();
+		try {
+			cliente = clienteService.findById(id);
+		} catch(DataAccessException e){
+			response.put("mensaje", "Error al realizar la consulta en la base de datos");
+			response.put("error", e.getMessage().concat(": ".concat(e.getMostSpecificCause().getMessage())));
+			return new ResponseEntity<Map<String,Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		if(cliente==null) {
+			response.put("mensaje", "El cliente ID: ".concat(id.toString().concat(" no existe en la base de datos.")));
+			return new ResponseEntity<Map<String,Object>>(response, HttpStatus.NOT_FOUND);
+		}
+		
+		return new ResponseEntity<Cliente>(cliente, HttpStatus.OK); 
+		
 	}
 	
 	@PostMapping("/clientes")
-	@ResponseStatus(HttpStatus.CREATED)
-	public Cliente create(@RequestBody Cliente cliente) {
-		return clienteService.save(cliente);
+	public ResponseEntity<?> create(@RequestBody Cliente cliente) {
+		
+		Cliente clienteNew= null;
+		Map<String,Object> response = new HashMap<>();
+		try {
+			clienteNew = clienteService.save(cliente);
+		} catch(DataAccessException e){
+			response.put("mensaje", "Error al realizar el registro en la base de datos");
+			response.put("error", e.getMessage().concat(": ".concat(e.getMostSpecificCause().getMessage())));
+			return new ResponseEntity<Map<String,Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		response.put("mensaje","El cliente ha sido creado con exito");
+		response.put("cliente", clienteNew );
+		return new ResponseEntity<Map<String,Object>>(response,(HttpStatus.CREATED));	
 	}
 	
 	@PutMapping("/clientes/{id}")
-	@ResponseStatus(HttpStatus.CREATED)
-	public Cliente update(@RequestBody Cliente cliente, @PathVariable Long id) {
+	public ResponseEntity<?> update(@RequestBody Cliente cliente, @PathVariable Long id) {
 		Cliente clienteActual = clienteService.findById(id);
+		Cliente clienteUpdated = null;
 		
-		clienteActual.setApellido(cliente.getApellido());
-		clienteActual.setNombre(cliente.getNombre());
-		clienteActual.setEmail(cliente.getEmail());
+		Map<String,Object> response = new HashMap<>();
 		
-		return clienteService.save(clienteActual);
+		if(clienteActual==null) {
+			response.put("mensaje", "Error: No se pudo editar el cliente con ID: ".concat(id.toString().concat(" en la base de datos.")));
+			return new ResponseEntity<Map<String,Object>>(response, HttpStatus.NOT_FOUND);
+		}
+		
+		try {
+			clienteActual.setApellido(cliente.getApellido());
+			clienteActual.setNombre(cliente.getNombre());
+			clienteActual.setEmail(cliente.getEmail());
+			clienteActual.setCreateAt(cliente.getCreateAt());
+			clienteUpdated = clienteService.save(clienteActual);
+		}catch(DataAccessException e){
+			response.put("mensaje", "Error al actualizar el registro en la base de datos");
+			response.put("error", e.getMessage().concat(": ".concat(e.getMostSpecificCause().getMessage())));
+			return new ResponseEntity<Map<String,Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		response.put("mensaje","El cliente ha sido actualizado con exito");
+		response.put("cliente", clienteUpdated );
+		
+		return new ResponseEntity<Map<String,Object>>(response,(HttpStatus.CREATED));
 	}
 	
 	@DeleteMapping("/clientes/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable Long id) {
-		clienteService.delete(id);
+	public ResponseEntity<?> delete(@PathVariable Long id) {
+		Map<String,Object> response = new HashMap<>();
+		Cliente clienteActual = clienteService.findById(id);
+		
+		try {
+			if(clienteActual==null) {
+				response.put("mensaje", "Error: No se pudo editar el cliente con ID: ".concat(id.toString().concat(" en la base de datos.")));
+				return new ResponseEntity<Map<String,Object>>(response, HttpStatus.NOT_FOUND);
+			}else {
+				clienteService.delete(id);
+			}
+		}catch(DataAccessException e){
+			response.put("mensaje", "Error al eliminar el registro en la base de datos");
+			response.put("error", e.getMessage().concat(": ".concat(e.getMostSpecificCause().getMessage())));
+			return new ResponseEntity<Map<String,Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		response.put("mensaje", "El ciente ha sido eliminado con exito!");
+		return new ResponseEntity<Map<String,Object>>(response, HttpStatus.OK);
 	}
 }
